refactor: Liner 클래스
