+++ src/main/java/domain/Liner.java
@@ -1,29 +1,25 @@
 package domain;
 
-import direction.Direction;
-import validator.HeightValidator;
-import validator.LadderNumberValidator;
 import validator.LinerValidator;
-import validator.SequenceValidator;
 
 public class Liner {
     private final int row;
     private final int numberOfPerson;
+    private final Ladder ladder;
 
     public Liner(int row, int numberOfPerson) {
         LinerValidator.validateLiner(row, numberOfPerson);
         this.row = row;
         this.numberOfPerson = numberOfPerson;
+        this.ladder = makeLadder();
     }
 
-    public void drawLine(int left, int right, int height, Row[] rows) {
-        LadderNumberValidator.validateLadderNumber(left, numberOfPerson);
-        LadderNumberValidator.validateLadderNumber(right, numberOfPerson);
-        SequenceValidator.validateSequence(left, right);
-        HeightValidator.validateHeight(left, right, rows, height);
+    private Ladder makeLadder() {
+        return new Ladder(row, numberOfPerson);
+    }
 
-        rows[height - 1].assignDirection(left - 1, Direction.RIGHT.getValue());
-        rows[height - 1].assignDirection(right - 1, Direction.LEFT.getValue());
+    public void drawLine(int left, int right, int height) {
+        ladder.drawLine(left, right, height, numberOfPerson);
     }
 
     public int getRow() {
@@ -34,4 +30,7 @@ public class Liner {
         return numberOfPerson;
     }
 
+    public Ladder getLadder() {
+        return ladder;
+    }
 }
